correccion: nuevo socio ingresa con la cuota y el pago generado

+++ app/Http/Controllers/PagoController.php
@@ -46,7 +46,7 @@ class PagoController extends ApiController
         try
         {
             $request->validate([
-                'cuotaDetalles' => 'required',
+                'cuotaDetalles' => 'nullable',
                 'fechaPago' => 'date|nullable'
             ]);
             
@@ -55,14 +55,16 @@ class PagoController extends ApiController
             $cuota = Cuota::whereId($cuota_id)->first();
             $service = new CuotaService();
 
-            //eliminamos los detalles de las cuotas y agregamos los que vienen en el request
-            $cuota->detalles()->delete();
-            //dd($cuota_detalles);
-            foreach ($cuota_detalles as $key => $cuota_detalle) {
-                $service->createCuotaDetalle($cuota->id, $cuota_detalle->cuota_detalle_tipo[0]->id, $cuota_detalle->monto, $cuota_detalle->descripcion);
+            if($cuota_detalles !== null) {
+                //eliminamos los detalles de las cuotas y agregamos los que vienen en el request
+                $cuota->detalles()->delete();
+                //dd($cuota_detalles);
+                foreach ($cuota_detalles as $key => $cuota_detalle) {
+                    $service->createCuotaDetalle($cuota->id, $cuota_detalle->cuota_detalle_tipo[0]->id, $cuota_detalle->monto, $cuota_detalle->descripcion);
 
-                if ($service->hasErrors()) {
-                    return $this->sendServiceError($service->getLastError());
+                    if ($service->hasErrors()) {
+                        return $this->sendServiceError($service->getLastError());
+                    }
                 }
             }
